NetworkPkg: Do not enforce secure RNG
Since edk2-stable202405 we require EFI_RNG_PROTOCOL for
various network stack drivers.

We can't avoid requiring the protocol, but we do not want to insist
that a secure algorithm is present.
If we do leave the Pcd TRUE, DxeNetLib logs at level DEBUG_ERROR when
using OVMF -device virtio-rng-pci (even though the RN generation
eventually succeeds), and it may do so with the available Rng in
various firmware too (and there, the RN generation may not succeed,
if none of the required algos are present).