Fix ASSERT() trigger at Ext4ReadFile() function
During fuzz testing, we encountered an ASSERT() condition
trigger in the Ext4ReadFile() function. This was because when
reading the file system superblock, the driver relied on the
s_rev_level field, based in which it initializes the Partition
structure. The case when (s_rev_level != EXT4_DYNAMIC_REV)
was handled incorrectly, leading to improver initialization
of the InodeSize field.

Additional checks have been added for this case.

Signed-off-by: Pavel Naberezhnev <pavelnaberezhnev@gmail.com>