MdePkg: Fix mismatched casts for PEI and DXE COPY_MEM and SET_MEM
The UEFI Platform Initialization Speciﬁcation and the Unified Extensible
Firmware Interface (UEFI) Specification (respectively for PEI and DXE)
determine that the signature for these methods must have a VOID return value,
but in all cases fields of the given types are initialised from methods
matching the CopyMem and SetMem declarations in BaseMemoryLib.h, which
have a VOID * return value.

There is no guarantee that a function returning a pointer can be safely
called by a caller expecting it to return void, even if this works in
practice with the calling conventions used, so wrapper methods are more
correct.

Fixing this is required to be able to compile with cast-function-type-mismatch
warning enabled for Xcode and clang toolchains, but it is at least arguably
a genuine unwanted mismatch in any case.

Signed-off-by: Mike Beaton <mjsbeaton@gmail.com>