.github/workflows: Change to non-recursive submodule checkout
Upstream, and our copy of the same file, recommend against
recursive submodule checkout:
https://github.com/mikebeaton/audk/blob/720bc76371ed5b57ce60226fb162d29c2e0b2c94/ReadMe.rst?plain=1#L274-L279

This is in order to avoid introducing unused dependencies,
i.e. no EDK2 build steps refer to any code in sub-submodules.

It is not believed that any code we have added needs
recursive submodule checkout either, so changing this:

 a) Removes unused dependencies
 b) Speeds up builds
 c) Makes builds less prone to network timeouts

Signed-off-by: Mike Beaton <mjsbeaton@gmail.com>